backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
